+++ ios/Classes/SwiftFlutterLoginVkPlugin.swift
@@ -9,7 +9,7 @@ enum PluginMethod: String {
 
 /// Arguments for method `PluginMethod.initSdk`
 enum InitSdkArg: String {
-    case appId, apiVersion, scope
+    case scope
 }
 
 /// Arguments for method `PluginMethod.logIn`
@@ -49,8 +49,8 @@ public class SwiftFlutterLoginVkPlugin: NSObject, FlutterPlugin {
             
             guard
                 let appId = Bundle.main.object(forInfoDictionaryKey: "VKAppId") as? String
-                else{
-                    result(FlutterError.invalidArgs("App id is invalid"))
+                else {
+                    result(FlutterError.invalidArgs("App id is not found. Please enter your VK application identifier in your Info.plist"))
                     return
             }